backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
chromadb>=0.4.0
python-dotenv>=1.0.0
pyyaml>=6.0
tiktoken>=0.5.0
sse-starlette>=1.6.0
